+++ pom.xml
@@ -8,13 +8,13 @@
         <version>7</version>
     </parent>
     <groupId>org.molgenis</groupId>
-    <artifactId>molgenis-genetics</artifactId>
+    <artifactId>molgenis-apps-core</artifactId>
     <version>1.0.0-SNAPSHOT</version>
     <packaging>pom</packaging>
 
-    <name>molgenis-genetics</name>
+    <name>>molgenis-apps-core</name>
     <description>
-        MOLGENIS Genetics is a repository containing modules and features for interacting with, searching through, and visualising genetic data
+        Molgenis core userinterface
     </description>
     <url>http://www.molgenis.org</url>
     <licenses>
@@ -23,8 +23,8 @@
         </license>
     </licenses>
     <scm>
-        <connection>scm:git:git://github.com/molgenis/molgenis-genetics.git</connection>
-        <url>https://github.com/molgenis/molgenis-genetics</url>
+        <connection>scm:git:git://github.com/molgenis/molgenis-apps-core.git</connection>
+        <url>https://github.com/molgenis/molgenis-apps-core</url>
     </scm>
 
     <properties>
@@ -32,15 +32,12 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <slf4j.version>1.7.12</slf4j.version>
         <logback.version>1.1.3</logback.version>
-
         <spring.version>4.3.7.RELEASE</spring.version>
         <spring.security.version>4.2.0.RELEASE</spring.security.version>
         <hibernate.validator.version>4.2.0.Final</hibernate.validator.version>
         <httpclient.version>4.5.1</httpclient.version>
         <google.auto-value.version>1.3</google.auto-value.version>
         <google.guava.version>21.0</google.guava.version>
-        <maven.build.timestamp.format>yyyy-MM-dd HH:mm z</maven.build.timestamp.format>
-        <molgenis.build.timestamp>${maven.build.timestamp}</molgenis.build.timestamp>
         <apache.rdf4j.version>2.0.2</apache.rdf4j.version>
     </properties>
 
@@ -400,7 +397,7 @@
     </dependencies>
 
     <modules>
-        <module>web-app</module>
+        <module>web-org.molgenis.org.controller.app</module>
     </modules>
 
 </project>
