+++ web-app/src/main/java/org/molgenis/app/WebAppSystemEntityRegistry.java
@@ -1,6 +1,6 @@
-package app;
+package org.molgenis.org.controller.app;
 
-import app.controller.HomeController;
+import org.molgenis.org.controller.app.controller.HomeController;
 import org.molgenis.auth.*;
 import org.molgenis.bootstrap.populate.SystemEntityRegistry;
 import org.molgenis.data.DataService;
