+++ web-app/src/main/java/org/molgenis/app/controller/ContactController.java
@@ -1,23 +1,23 @@
-package app.controller;
+package org.molgenis.org.controller.app.controller;
 
 import org.molgenis.ui.MolgenisPluginController;
 import org.molgenis.ui.controller.AbstractStaticContentController;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import static app.controller.NewsController.URI;
+import static org.molgenis.org.controller.app.controller.ContactController.URI;
 
 /**
- * Controller that handles news page requests
+ * Controller that handles contact page requests
  */
 @Controller
 @RequestMapping(URI)
-public class NewsController extends AbstractStaticContentController
+public class ContactController extends AbstractStaticContentController
 {
-	public static final String ID = "news";
+	public static final String ID = "contact";
 	public static final String URI = MolgenisPluginController.PLUGIN_URI_PREFIX + ID;
 
-	public NewsController()
+	public ContactController()
 	{
 		super(ID, URI);
 	}
