+++ web-app/src/main/java/org/molgenis/app/controller/NewsController.java
@@ -1,24 +1,23 @@
-package app.controller;
+package org.molgenis.org.controller.app.controller;
 
 import org.molgenis.ui.MolgenisPluginController;
 import org.molgenis.ui.controller.AbstractStaticContentController;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import static app.controller.ReferencesController.URI;
+import static org.molgenis.org.controller.app.controller.NewsController.URI;
 
 /**
- * Controller that handles references page requests
+ * Controller that handles news page requests
  */
 @Controller
 @RequestMapping(URI)
-public class ReferencesController extends AbstractStaticContentController
+public class NewsController extends AbstractStaticContentController
 {
-	public static final String ID = "references";
+	public static final String ID = "news";
 	public static final String URI = MolgenisPluginController.PLUGIN_URI_PREFIX + ID;
 
-	public ReferencesController()
-
+	public NewsController()
 	{
 		super(ID, URI);
 	}
