+++ web-app/src/main/resources/logback.xml
@@ -1,0 +1,15 @@
+<configuration>
+	<appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+		<encoder>
+			<pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+		</encoder>
+	</appender>
+	<appender name="jobExecutionLogAppender" class="org.molgenis.data.jobs.JobExecutionLogAppender"/>
+	<logger name="org.molgenis" level="INFO"/>
+	<root level="WARN">
+		<appender-ref ref="STDOUT" />
+	</root>
+    <logger name="org.molgenis.data.jobs.model.JobExecution" level="ALL">
+        <appender-ref ref="jobExecutionLogAppender"/>
+    </logger>
+</configuration>
