+++ web-app/src/main/resources/molgenis.properties
@@ -1,0 +1,6 @@
+###############################################################
+# MOLGENIS configuration file                                 #
+###############################################################
+
+molgenis.version=${pom.version}
+molgenis.build.date=${molgenis.build.timestamp}
