+++ web-app/src/main/resources/molgenis_ui.xml
@@ -1,0 +1,44 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<molgenis label="OMX" name="omx">
+	<menu label="Home" name="main">
+		<plugin id="home" name="Home" url="home"/>
+        <plugin id="importwizard" name="Upload" url="importwizard"/>
+		<plugin id="dataexplorer" name="Data Explorer" url="dataexplorer"/>
+		<plugin id="catalogue" name="Catalogue" url="catalogue"/>
+		<menu label="Data Integration" name="dataintegration">
+			<plugin id="mappingservice" name="Mapping Service" url="mappingservice"/>
+			<plugin id="sortaservice" name="SORTA" url="sortaservice"/>
+			<plugin id="tagwizard" name="Tag Wizard" url="tagwizard"/>
+            <plugin id="ontologymanager" name="Ontology manager" url="ontologymanager"/>
+		</menu>
+        <menu label="Plugins" name="plugins">
+            <plugin id="apps" name="Apps" url="apps"/>
+            <plugin id="questionnaires" name="Questionnaires" url="questionnaires"/>
+        	<plugin id="pathways" name="Pathways" url="pathways"/>
+            <plugin id="standardsregistry" name="Model registry" url="standardsregistry" />
+            <plugin id="background" name="Background" url="background"/>
+            <plugin id="contact" name="Contact" url="contact"/>
+            <plugin id="news" name="News" url="news"/>
+            <plugin id="references" name="References" url="references"/>
+            <plugin id="scripts" name="Scripts" url="scripts" />
+            <plugin id="jobs" name="Jobs" url="jobs" />
+            <plugin id="fileingest" name="Automatic imports" url="fileingest" />
+            <plugin id="gavin-org.molgenis.org.controller.app" name="Gavin" url="gavin-org.molgenis.org.controller.app"/>
+            <plugin id="genenetwork" name="Gene network" url="genenetwork"/>
+            <plugin id="genetics-diag-org.molgenis.org.controller.app" name="Genetics diagnostics" url="genetics-diag-org.molgenis.org.controller.app"/>
+	    </menu>
+        <menu label="Admin" name="admin">
+	        <plugin id="usermanager" name="User Manager" url="usermanager"/>
+	        <plugin id="permissionmanager" name="Permission Manager" url="permissionmanager"/>
+        	<plugin id="menumanager" name="Menu Manager" url="menumanager"/>
+        	<plugin id="thememanager" name="Theme Manager" url="thememanager"/>
+        	<plugin id="indexmanager" name="Index manager" url="indexmanager"/>
+        	<plugin id="logmanager" name="Log manager" url="logmanager"/>
+        	<plugin id="settingsmanager" name="Settings" url="settingsmanager"/>
+		</menu>
+		<menu label="About" name="background">
+            <plugin id="feedback" name="Feedback" url="feedback"/>
+        </menu>
+		<plugin id="useraccount" name="Account" url="useraccount"/>
+	</menu>
+</molgenis>
